requirement.txt
dj-database-url==0.4.2
Django==1.9
django-admin-bootstrapped-compatible==3.1.7
django-crispy-forms==1.6.1
pytz==2017.2
whitenoise==3.3.1
